# docs/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateLineComponentDto": {
        "properties": {
          "code": {
            "maxLength": 40,
            "pattern": "^[A-Z0-9][A-Z0-9_-]*$",
            "type": "string"
          },
          "componentTypeId": { "format": "uuid", "type": "string" },
          "displayOrder": { "minimum": 1, "type": "number" },
          "name": { "maxLength": 120, "minLength": 1, "type": "string" }
        },
        "required": ["code", "name", "displayOrder", "componentTypeId"],
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": { "example": "AUTHORIZATION_DENIED", "type": "string" },
          "correlationId": { "format": "uuid", "type": "string" },
          "message": {
            "oneOf": [
              { "type": "string" },
              { "items": { "type": "string" }, "type": "array" }
            ]
          },
          "path": {
            "example": "/api/v1/organizations/30000000-0000-4000-8000-000000000001/plants",
            "type": "string"
          },
          "statusCode": { "example": 403, "type": "integer" },
          "timestamp": { "format": "date-time", "type": "string" }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "path",
          "timestamp",
          "correlationId"
        ],
        "type": "object"
      },
...
      "Object": { "properties": {}, "type": "object" },
...
      "PageCatalog": {
        "properties": {
          "items": {
            "items": { "$ref": "#/components/schemas/CatalogItem" },
            "type": "array"
          },
          "page": { "minimum": 1, "type": "integer" },
          "pageSize": { "maximum": 100, "minimum": 1, "type": "integer" },
          "total": { "minimum": 0, "type": "integer" },
          "totalPages": { "minimum": 0, "type": "integer" }
        },
        "required": ["items", "page", "pageSize", "total", "totalPages"],
        "type": "object"
      },
...
      "SetCatalogStateDto": {
        "properties": { "active": { "type": "boolean" } },
        "required": ["active"],
        "type": "object"
      },
...
      "UpdateLineComponentDto": {
        "properties": {
          "code": {
            "maxLength": 40,
            "pattern": "^[A-Z0-9][A-Z0-9_-]*$",
            "type": "string"
          },
          "componentTypeId": { "format": "uuid", "type": "string" },
          "displayOrder": { "minimum": 1, "type": "number" },
          "name": { "maxLength": 120, "minLength": 1, "type": "string" }
        },
        "type": "object"
      },
...
      "UpdatePlantDto": {
        "properties": {
          "code": {
            "maxLength": 40,
            "pattern": "^[A-Z0-9][A-Z0-9_-]*$",
            "type": "string"
          },
          "name": { "maxLength": 120, "minLength": 1, "type": "string" },
          "timezone": {
            "default": "America/Costa_Rica",
            "maxLength": 80,
            "minLength": 1,
            "type": "object"
          }
        },
        "type": "object"
      },
      "UpdateProductionLineDto": {
        "properties": {
          "code": {
            "maxLength": 40,
            "pattern": "^[A-Z0-9][A-Z0-9_-]*$",
            "type": "string"
          },
          "displayOrder": { "minimum": 1, "type": "number" },
          "name": { "maxLength": 120, "minLength": 1, "type": "string" }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "supabase": {
        "bearerFormat": "Supabase access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Contrato backend del sistema de planta. Todos los ejemplos son ficticios.",
    "title": "Industrias Doradas API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations/{organizationId}/component-types": {
      "get": {
        "operationId": "CatalogsController_listComponentTypes",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PageCatalog" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Catalogs"]
      }
    },
    "/api/v1/organizations/{organizationId}/components/{id}": {
      "patch": {
        "operationId": "CatalogsController_updateComponent",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLineComponentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "object" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Catalogs"]
      }
    },
    "/api/v1/organizations/{organizationId}/components/{id}/state": {
      "patch": {
        "operationId": "CatalogsController_setComponentState",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SetCatalogStateDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "object" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Catalogs"]
      }
    },
    "/api/v1/organizations/{organizationId}/lines/{id}": {
      "patch": {
        "operationId": "CatalogsController_updateLine",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductionLineDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "object" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Catalogs"]
      }
    },
    "/api/v1/organizations/{organizationId}/lines/{id}/state": {
      "patch": {
        "operationId": "CatalogsController_setLineState",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SetCatalogStateDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "object" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Catalogs"]
      }
    },
    "/api/v1/organizations/{organizationId}/lines/{lineId}/components": {
      "get": {
        "operationId": "CatalogsController_listComponents",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "lineId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "page",
            "required": true,
            "schema": {
              "allOf": [{ "$ref": "#/components/schemas/Object" }],
              "default": 1,
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": true,
            "schema": {
              "allOf": [{ "$ref": "#/components/schemas/Object" }],
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": { "maxLength": 100, "type": "string" }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "default": "active",
              "enum": ["all", "active", "inactive"],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PageCatalog" }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Catalogs"]
      },
      "post": {
        "operationId": "CatalogsController_createComponent",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "lineId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLineComponentDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": { "schema": { "type": "object" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Catalogs"]
      }
    },
...
    "/api/v1/organizations/{organizationId}/plants/{id}": {
      "patch": {
        "operationId": "CatalogsController_updatePlant",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePlantDto" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "object" } }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            },
            "description": "Error estable"
          }
        },
        "security": [{ "supabase": [] }],
        "tags": ["Catalogs"]
      }
    },
...
  },
  "servers": [],
...
}